Moved more metadata to implicit parameters
Also fixed a lot of tests that weren't passing after the recent refactoring.

+++ src/main/scala/eusocialcooperation/scheduler/Worker.scala
@@ -71,7 +71,7 @@ object Worker {
       val loopDelayMs = {
         // TODO: There's a concept in typesafe config of reference.conf, I think, that handles defaults more simply than this, but idk
         if (config.hasPath("loopDelayMs")) {
-          config.getInt("loopDelayMs").toLong
+          config.getLong("loopDelayMs").toLong
         } else {
           Worker.defaultLoopDelayMs
         }
